+++ hss_okta_flutter/example/lib/screens/home_screen.dart
@@ -141,7 +141,14 @@ class _HomeScreenState extends State<HomeScreen> {
                           options: AuthorizeOptions(
                         responseType: ['token', 'id_token'],
                       ));
-                  debugPrint(value.tokens.accessToken?.accessToken ?? '');
+
+                  if (value.tokens.accessToken?.accessToken != null) {
+                    if (mounted) {
+                      ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+                          content: Text(
+                              'Access Token: ${value.tokens.accessToken?.accessToken}')));
+                    }
+                  }
                 } else {
                   var result = await Navigator.of(formContext).push(
                       MaterialPageRoute(
@@ -271,8 +278,10 @@ class _HomeScreenState extends State<HomeScreen> {
                       _pageController.jumpTo(0);
                     }
                   } catch (e) {
-                    ScaffoldMessenger.of(profileContext).showSnackBar(
-                        SnackBar(content: Text('Error: ${e.toString()}')));
+                    if (mounted) {
+                      ScaffoldMessenger.of(profileContext).showSnackBar(
+                          SnackBar(content: Text('Error: ${e.toString()}')));
+                    }
                   }
                 },
                 child: const Text('WebAuth Signout')),
@@ -291,8 +300,10 @@ class _HomeScreenState extends State<HomeScreen> {
                         }
                       });
                     } catch (e) {
-                      ScaffoldMessenger.of(profileContext).showSnackBar(
-                          SnackBar(content: Text('Error: ${e.toString()}')));
+                      if (mounted) {
+                        ScaffoldMessenger.of(profileContext).showSnackBar(
+                            SnackBar(content: Text('Error: ${e.toString()}')));
+                      }
                     }
                   }
                 },