+++ hss_okta_flutter/CHANGELOG.md
@@ -1,11 +1,12 @@
 # 1.2.2
 * Made OktaAuth private
-* Made OktaConfig Private
 * Added getter for OktaAuth
-* Added getter for OktaConfig
+* Removed OktaConfig
 * Added AuthStateManager
 * Added TokenManager
 * Added Helper methods for OktaAuth
+* Changed initialize auth client parameters to OktaConfig
+* Added additional okta config parameters
 
 # 1.1.1
 * Updated readme.md